Separated Classes
Separated the classes and added block quotes. Replaced some constant
values with constant variables.
